Refactor the native library interop code
The GmicSharp native libraries now have the same name but are located
in different folders.
This simplifies making the GmicSharp work on .NET Core and non-Windows
operating systems, it also removes the need to have separate P/Invoke
definitions for the x64 and x86 versions of the native libraries.

The appropriate libGmicSharpNative version will be loaded into the
process at runtime, this allows custom library locations to be used
with both the .NET Framework and .NET Core.