Replace the DllImport pre-load trick with delegates
It appears that loading the DLL into the process from a custom path
before calling it via DllImport does not work with .NET Core 3.1 on
Ubuntu 20.4.
The DLL is loaded into the process, but a DllNotFoundException is
thrown when any method is called using DllImport.

Using GetDelegateForFunctionPointer to retrieve a delegate for each
required export works correctly on both Linux and Windows.