Allow G'MIC instances to be run in parallel
All of the RunGmic* methods in the Gmic<TGmicBitmap> class now take the
input images as a parameter, this allows the RunGmicAsync methods to
execute multiple independent instances of the G'MIC interpreter.

The Gmic<TGmicBitmap> class no longer implements IDisposable or the
Event-based Asynchronous Pattern.
Removed the AddInputImage, ClearInputImages, void RunGmicAsync(string)
and RunGmicAsyncCancel methods.
Removed the RunGmicCompleted and RunGmicProgressChanged events.
Renamed all of the RunGmicTaskAsync methods to RunGmicAsync.

Added a Name property to the GmicBitmap class, this replaces the
AddInputImage(TGmicBitmap, string) method in the Gmic<TGmicBitmap>
class.